move containter services tests to separate directory
