+++ tests_integration/tests_using_container_services/spcs/docker/build_and_push_all.sh
