+++ tests_integration/tests_using_container_services/spcs/docker/echo_service/Dockerfile
