+++ tests_integration/tests_using_container_services/spcs/docker/test_counter/build_and_push.sh
