+++ tests_integration/tests_using_container_services/spcs/docker/test_counter/simple_counter.py
