+++ tests_integration/tests_using_container_services/spcs/test_registry.py
