+++ tests_integration/tests_using_container_services/spcs/testing_utils/compute_pool_utils.py
