+++ tests_integration/tests_using_container_services/spcs/testing_utils/image_repository_utils.py
